front: move split points in a separate file
